+++ src/aio.c
@@ -671,9 +671,10 @@ SEXP rnng_aio_http(SEXP aio) {
     REprintf("HTTP Server Response: %d %s\n", code, nng_http_res_get_reason(handle->res));
     if (code >= 300 && code < 400) {
       SEXP out;
-      PROTECT(out = Rf_allocVector(VECSXP, 2));
+      PROTECT(out = Rf_allocVector(VECSXP, 3));
       SET_VECTOR_ELT(out, 0, Rf_ScalarInteger(code));
-      SET_VECTOR_ELT(out, 1, Rf_mkString(nng_http_res_get_header(handle->res, "Location")));
+      SET_VECTOR_ELT(out, 1, Rf_mkString(nng_http_res_get_header(handle->res, "Date")));
+      SET_VECTOR_ELT(out, 2, Rf_mkString(nng_http_res_get_header(handle->res, "Location")));
       UNPROTECT(1);
       return out;
     }
@@ -683,12 +684,13 @@ SEXP rnng_aio_http(SEXP aio) {
   size_t sz;
   SEXP out, vec;
 
-  PROTECT(out = Rf_allocVector(VECSXP, 2));
+  PROTECT(out = Rf_allocVector(VECSXP, 3));
   SET_VECTOR_ELT(out, 0, Rf_ScalarInteger(code));
+  SET_VECTOR_ELT(out, 1, Rf_mkString(nng_http_res_get_header(handle->res, "Date")));
   nng_http_res_get_data(handle->res, &dat, &sz);
   vec = Rf_allocVector(RAWSXP, sz);
   memcpy(RAW(vec), dat, sz);
-  SET_VECTOR_ELT(out, 1, vec);
+  SET_VECTOR_ELT(out, 2, vec);
 
   UNPROTECT(1);
   return out;